requirements.txt
torch
torchaudio
transformers
faiss-cpu
pydub
numpy